requirements.txt
openai==1.5.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.26.0
pydantic==2.4.2
tqdm==4.66.1